move molmoact2 config logic into config
